pyproject.toml
[project]
name = "dse855-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "zabbix-utils[async]>=2.0.2",
]
